requirements.txt
# MCP Server Framework
fastmcp==2.14.1

# Google API
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.35.0
python-multipart==0.0.18

# Pydantic for settings and validation
pydantic>=2.11.7
pydantic-settings>=2.6.1

# Database (optional for future use)
sqlalchemy==2.0.36
alembic==1.14.0

# Caching
redis==5.2.0
hiredis==3.0.0

# Utilities
python-dotenv==1.1.0
httpx==0.28.1
tenacity==9.0.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.4
mypy==1.13.0
